Make CLI only depend on the new public API
